containers refactor #1774
