+++ src/js/containers/mainModal/RecentProjectsContainer.js
@@ -4,12 +4,12 @@ import path from 'path-extra';
 import fs from 'fs-extra';
 import { Glyphicon } from 'react-bootstrap/lib';
 // components
-import RecentProjects from '../components/RecentProjects';
+import RecentProjects from '../../components/RecentProjects';
 // actions
-import * as recentProjectsActions from '../actions/RecentProjectsActions.js';
-import * as ModalActions from '../actions/ModalActions.js';
-import * as AlertModalActions from '../actions/AlertModalActions.js';
-import { openProject } from '../actions/GetDataActions.js';
+import * as recentProjectsActions from '../../actions/RecentProjectsActions.js';
+import * as ModalActions from '../../actions/ModalActions.js';
+import * as AlertModalActions from '../../actions/AlertModalActions.js';
+import { openProject } from '../../actions/GetDataActions.js';
 // constant declaration
 const DEFAULT_SAVE = path.join(path.homedir(), 'translationCore');
